messages refactor
