+++ scripts/kong_data/requests.sh
@@ -1,17 +1,11 @@
 #!/usr/bin/env bash
 
-if [ -z "$1" ]
-	then
-		NETWORK=""
-	else
-		NETWORK="--network $1"
-fi
+NETWORK="--network ic"
 IDENTITY="--identity kong_user1"
-
 KONG_DATA=$(dfx canister id ${NETWORK} kong_data)
 
-REQUEST_ID='opt 8'
-NUM_REQUESTS='opt 10'
+request_id="opt 8"
+num_requests="opt 10"
 
 #dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} requests --output json "(null, null)" | jq
-dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} requests --output json "($REQUEST_ID, $NUM_REQUESTS)" | jq
+dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} requests --output json "($request_id, $num_requests)" | jq
