+++ scripts/kong_data/txs.sh
@@ -1,19 +1,13 @@
 #!/usr/bin/env bash
 
-if [ -z "$1" ]
-	then
-		NETWORK=""
-	else
-		NETWORK="--network $1"
-fi
+NETWORK="--network ic"
 IDENTITY="--identity kong_user1"
-
 KONG_DATA=$(dfx canister id ${NETWORK} kong_data)
 
-MY_TXS='opt false'
-TX_ID='null'
-TOKEN_ID='null'
-NUM_TXS='opt 10'
+my_txs="opt false"
+tx_id="null"
+token_id="null"
+num_txs="opt 10"
 
-dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} txs --output json "(null, null, null, null)" | jq
-#dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} txs --output json "($MY_TXS, $TX_ID, $TOKEN_ID, $NUM_TXS)" | jq
+#dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} txs --output json "(null, null, null, null)" | jq
+dfx canister call ${NETWORK} ${IDENTITY} ${KONG_DATA} txs --output json "($my_txs, $tx_id, $token_id, $num_txs)" | jq
