+++ src/kong_backend/src/messages/messages.rs
@@ -1,16 +1,21 @@
 use ic_cdk::query;
 
-use super::message_reply::{to_messages_reply, MessagesReply};
+use super::message_reply::MessagesReply;
+use super::message_reply_helpers::to_messages_reply;
 
-use crate::ic::guards::not_in_maintenance_mode;
+use crate::ic::guards::not_in_maintenance_mode_and_caller_is_not_anonymous;
 use crate::stable_message::message_map;
+use crate::stable_user::user_map;
 
-const MAX_MESSAGES: usize = 10;
-
-#[query(guard = "not_in_maintenance_mode")]
+#[query(guard = "not_in_maintenance_mode_and_caller_is_not_anonymous")]
 pub async fn messages(message_id: Option<u64>) -> Result<Vec<MessagesReply>, String> {
-    Ok(match message_id {
-        Some(message_id) => message_map::get_by_message_id(message_id).iter().map(to_messages_reply).collect(),
-        None => message_map::get(MAX_MESSAGES).iter().map(to_messages_reply).collect(),
-    })
+    let user_id = match user_map::get_by_caller() {
+        Ok(Some(caller)) => caller.user_id,
+        Ok(None) | Err(_) => return Ok(Vec::new()),
+    };
+    let messages = message_map::get_by_message_id(message_id, Some(user_id), None)
+        .iter()
+        .map(to_messages_reply)
+        .collect();
+    Ok(messages)
 }
