+++ src/kong_backend/src/messages/mod.rs
@@ -1,3 +1,4 @@
 pub mod message_reply;
+pub mod message_reply_helpers;
 #[allow(clippy::module_inception)]
 pub mod messages;
