+++ src/kong_backend/src/stable_message/message_map.rs
@@ -1,36 +1,29 @@
+use std::ops::Bound;
+
 use super::stable_message::{StableMessage, StableMessageId};
 
 use crate::stable_kong_settings::kong_settings;
 use crate::stable_memory::MESSAGE_MAP;
-use crate::stable_user::stable_user::ALL_USERS_USER_ID;
-use crate::stable_user::user_map;
 
-pub fn get_by_message_id(message_id: u64) -> Option<StableMessage> {
-    let user_id = user_map::get_by_caller().ok().flatten()?.user_id;
-    MESSAGE_MAP.with(|m| {
-        m.borrow().get(&StableMessageId(message_id)).and_then(|v| {
-            if v.to_user_id == user_id || v.to_user_id == ALL_USERS_USER_ID {
-                return Some(v);
-            }
-            None
-        })
-    })
-}
+const MAX_MESSAGES: usize = 10;
 
-pub fn get(num_messages: usize) -> Vec<StableMessage> {
-    let user_id = match user_map::get_by_caller() {
-        Ok(Some(caller)) => caller.user_id,
-        Ok(None) | Err(_) => return Vec::new(),
-    };
+pub fn get_by_message_id(start_message_id: Option<u64>, user_id: Option<u32>, num_messages: Option<usize>) -> Vec<StableMessage> {
     MESSAGE_MAP.with(|m| {
-        m.borrow()
-            .iter()
+        let map = m.borrow();
+        let start_message_id = start_message_id.unwrap_or(map.last_key_value().map_or(0, |(k, _)| k.0));
+        let num_messages = match num_messages {
+            Some(num_messages) => std::cmp::min(num_messages, MAX_MESSAGES),
+            None => 1,
+        };
+        map.range((Bound::Unbounded, Bound::Included(&StableMessageId(start_message_id))))
             .rev()
             .filter_map(|(_, v)| {
-                if v.to_user_id == user_id || v.to_user_id == ALL_USERS_USER_ID {
-                    return Some(v.clone());
+                if let Some(user_id) = user_id {
+                    if v.to_user_id != user_id {
+                        return None;
+                    }
                 }
-                None
+                Some(v)
             })
             .take(num_messages)
             .collect()
