+++ src/kong_backend/src/stable_pool/check_token_balance.rs
@@ -42,7 +42,8 @@ pub async fn check_token_balance(token: &StableToken) -> Result<(StableToken, Na
     };
     // iterate over all pools and sum up the balances
     POOL_MAP.with(|m| {
-        for (_, v) in m.borrow().iter() {
+        let map = m.borrow();
+        for (_, v) in map.iter() {
             if v.token_0().token_id() == token_id {
                 // expected_balance += v.balance_0 + v.lp_fee_0 + v.kong_fee_0;
                 expected_balance.balance += nat_add(&nat_add(&v.balance_0, &v.lp_fee_0), &v.kong_fee_0);
@@ -66,7 +67,8 @@ pub async fn check_token_balance(token: &StableToken) -> Result<(StableToken, Na
     });
     // add unclaimed claims back to balances
     CLAIM_MAP.with(|m| {
-        for (_, v) in m.borrow().iter() {
+        let map = m.borrow();
+        for (_, v) in map.iter() {
             if stable_claim_helpers::get_token(&v).token_id() == token_id && v.status == ClaimStatus::Unclaimed {
                 expected_balance.unclaimed_claims += v.amount
             }
